refactor: decompose DonationPage into smaller components and add types
