+++ package.json
@@ -11,6 +11,7 @@
   "license": "MIT",
   "devDependencies": {
     "@types/react": "^18.2.46",
+    "@types/react-copy-to-clipboard": "^5.0.7",
     "@types/react-dom": "^18.2.18",
     "@typescript-eslint/eslint-plugin": "^6.16.0",
     "@typescript-eslint/parser": "^6.16.0",
